+++ src/assets/css/style.css
@@ -178,7 +178,8 @@ dialog::backdrop {
 }
 
 .comment-container{
-  width: 80%;
+  width: 100%;
+  height: 20vh;
 }
 @media (min-width: 768px) {
   .hamburger-menu {
@@ -239,5 +240,10 @@ dialog::backdrop {
   .comment-modal ::backdrop {
     background-color: var(--clr-primary);
     opacity: 0.2;
+  }
+
+  .comment-container{
+    width: 100%;
+    height: 50vh;
   }
 }
