+++ src/modules/Card.js
@@ -1,5 +1,4 @@
 /* eslint-disable no-unused-vars */
-
 import Involvement from "./Involvement.js";
 
 export default class {
@@ -7,6 +6,8 @@ export default class {
     this.imageUrl = '';
     this.likes = 0;
     this.card = this.#createCardTemplate();
+    this.comments = 0;
+    this.commentsArr = []
   }
 
   async fetchImage() {
@@ -70,19 +71,25 @@ export default class {
 
   async displayComment(commentModal, index, appId) {
     const commentBtn = this.card.querySelector('.btn-comment');
-    const closeComment = document.querySelector('.close-modal');
+    const closeComment = commentModal.querySelector('.close-modal');
     const involvement = new Involvement();
 
     commentBtn.addEventListener('click', async () => {
       let comments = await involvement.getComments(index, appId);
+      if (!comments.error){ 
+        this.commentsArr.push(comments)
+        console.log(this.commentsArr)
+        this.comments = [...comments].length 
+      }
       commentModal.children[1].innerHTML =`
           <div class="comment-container card-img-container">
             <img class="card-img" src="${this.imageUrl}" alt="">
           </div>
           <h2 class="card-details-header">Avocado Salad</h2>
-           <h3 class="card-details-header clr-primary">Comments</h3>
+           <h3 class="card-details-header clr-primary">Comments (${this.comments || 0})</h3>
       `;
       if (!comments.error){
+        const ul = document.createElement('ul');
         [...comments].forEach((comment)=>{
           ul.innerHTML +=   `<li> ${comment.creation_date}: ${comment.username} - ${comment.comment} </li>`
         })
@@ -94,5 +101,32 @@ export default class {
     closeComment.addEventListener('click', () => {
       commentModal.close();
     });
+  }
+
+  async addComment(commentModal,){
+    let commentData = {};
+    const commentInputs = commentModal.querySelectorAll('.form-input');
+    const submitComment = commentModal.querySelector('#commentBtn');
+
+
+    commentInputs.forEach((input)=> {
+      input.addEventListener('change', (e)=>{
+          if (input.name === 'name'){
+            commentData.name = e.target.value;
+          }
+          if (input.name === 'comment'){
+            commentData.comment = e.target.value;
+          }
+        submitComment.value = JSON.stringify(commentData);
+      })
+    })
+    commentModal.addEventListener('close', async (e)=> {
+      if (commentModal.returnValue){
+        const formData = JSON.parse(commentModal.returnValue)
+        if(formData.name){
+        await involvement.createComment(formData.id, app.id, formData.name, formData.comment)
+      }
+    }
+  })
   }
 }
